feat: migrate to sink actions for pattern handlers
